Bug fixes and format
Fixed and reworked how the format function works. While it seems to be
bug free, it is not quite finished yet. However, the cases which are not
finished are not ones I typically come across. So it seems to be OK for
now. Bug fixes for other functions have been implemented as well.